Hotdog-not-hotdog/requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.26.4
azure-functions==1.14.0
azure-storage-blob==12.16.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
pycparser==2.21
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.5.0
urllib3==2.0.2
